add progrres bar and fix adding new post

+++ app/src/main/java/com/example/belablok/model/Post.kt
@@ -2,7 +2,7 @@ package com.example.belablok.model
 
 
 data class Post(
-    val id: Int,
+    val id: Float,
     val title: String?,
     val imageUrl: String
 )
