+++ app/src/main/java/com/example/belablok/ui/viewmodels/DeletePostViewModel.kt
@@ -6,9 +6,10 @@ import com.example.belablok.repositories.PostsRepository
 import kotlinx.coroutines.launch
 
 class DeletePostViewModel(private val postsRepository: PostsRepository) : ViewModel() {
-    fun deletePost(id: Int) {
+    fun deletePost(id: Float) {
         viewModelScope.launch {
             postsRepository.deletePost(id)
         }
     }
+
 }
